+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -134,7 +134,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-02T10:53:00.225Z",
+    "2026-02-03T05:16:03.721Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {