+++ packages/orq-rc/RELEASES.md
@@ -6329,3 +6329,13 @@ Based on:
 - [typescript v4.3.0-rc.17] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.17 - packages/orq-rc
+
+## 2026-02-03 05:18:56
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.700.2 (2.801.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.18] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.18] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.18 - packages/orq-rc